renamed DatesUtils.ts to date-utils.ts
